Updating esxi collector to publish values of mor references
Given that for the mast majority of stats exposed via the vSphere python sdk
we want to get the average, we're going to publish ALL averages via statsd

We're mostly doing this as a diagnostic to figure out which values we
actually care about.

There's some complexity here, because we only want to make one http request
for the complete list of more references...once we get the values,  we inspect
the elements to group them according to group and counter.

Returns a tuple of (key, value) where key is formatted "key(unit)x<len_stats>"
and the value is the average of all strings in the list casted to floats.

The approximate time is still ~10 seconds, so I've making the requests in
parallel.